requirements.txt
numpy
pandas
scikit-learn
scipy
psycopg2-binary
flask
streamlit 
matplotlib
seaborn
plotly